fix test on 32 bits

+++ tests/unit-fail/incompleteStruct-link2.c
@@ -2,5 +2,5 @@
 struct foo { int y; } x;
 void bar(struct foo *p) {
   int *p2 = (int *)p;
-  long *p3 = (long *)p;
+  long long *p3 = (long long *)p;
 }
